+++ packages/veui/src/components/Tabs/Tab.vue
@@ -8,6 +8,7 @@
 import { pick, find, uniqueId } from 'lodash'
 import { getTypedAncestorTracker } from '../../utils/helper'
 import { getIndexOfType } from '../../utils/context'
+import '../../config/uiTypes'
 
 export default {
   name: 'veui-tab',